br/utils: add nil storage panic check in keyspace mode
